Program portable over ten Scheme implementations.
This includes a local test runner, github action continuation
integration, dockerfile, and script to install on a Debian several
Scheme implementation, see `./venv scheme-live available`.

It demonstrate how to build a program that rely on a library that is
portable to several implementations. Here is what the repository looks
like:

% tree .
.
├── gerbil.pkg
├── live
│   ├── hello
│   │   └── body.scm
│   ├── hello.chez.sls
│   ├── hello.guile.sld
│   ├── hello.rkt
│   └── hello.sld

Where `body.scm` contains the code. The other files are importing the
required libraries, which may be r7rs, r6rs or something else entirely
depending on what each implementation decide to use as an entry point
into the library... It works with racket, see `hello.rkt`.

ref: https://github.com/lexi-lambda/racket-r7rs/issues/16

The directory ./scheme/ contains r7rs shims for chez, to make it
possible to run ./local/bin/hello-schemer and
./local/bin/hello-schemer-star.

Here is the content of `hello-schemer`:

```scheme
 #!/usr/bin/env -S scheme-live current run
 (import (scheme base))
 (import (scheme write))

 (display "Hello schemer!")
 (newline)
```

The bang in the first line is mostly useless. It was meant to make it
possible to run `hello-schemer` without going through the woops of
spelling the full:

  scheme-live fantastic run ./local/bin/hello-schemer

That dream was not completly realised, and calling the following:

  ./venv hello-schemer

Result in the following notice:

  Retry with something like:

    SCHEME_LIVE_CURRENT=fantastic hello-schemer

  Or:

    export SCHEME_LIVE_CURRENT=fantastic

  Then call the program.

That is to say, the user needs to set SCHEME_LIVE_CURRENT to something
that is supported.

At somepoint, we may get rid of the shell code, we are not there yet.